FEAT/PIN-2514

Closes PIN-2512 and PIN-2489

See merge request humansis/mobile-apps/user-app!130

+++ app/src/main/java/cz/applifting/humansis/model/api/Project.kt
@@ -11,7 +11,7 @@ data class Project(
     @SerializedName("start_date") val startDate: String?,
     @SerializedName("end_date") val endDate: String?,
     @SerializedName("number_of_households") val numberOfHouseholds: Int?,
-    val target: Int?,
+    val target: Long?,
     val notes: String?,
     val iso3: String?,
     val donors: List<Any>?,
